% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{am_rollback}
\alias{am_rollback}
\title{Roll back pending operations}
\usage{
am_rollback(doc)
}
\arguments{
\item{doc}{An Automerge document}
}
\value{
The document \code{doc} (invisibly)
}
\description{
Cancels all pending operations in the current transaction without
committing them. This allows you to discard changes since the last
commit.
}
\examples{
doc <- am_create()

am_put(doc, AM_ROOT, "key", "value")
# Changed my mind, discard the put
am_rollback(doc)

am_close(doc)

}
