% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.list.am_list}
\alias{as.list.am_list}
\title{Convert Automerge list to R list}
\usage{
\method{as.list}{am_list}(x, doc = NULL, ...)
}
\arguments{
\item{x}{An Automerge list object}

\item{doc}{The document containing this object (automatically provided)}

\item{...}{Additional arguments (unused)}
}
\value{
Unnamed list
}
\description{
Recursively converts an Automerge list to an unnamed R list.
}
\keyword{internal}
