% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.list.am_map}
\alias{as.list.am_map}
\title{Convert Automerge map to R list}
\usage{
\method{as.list}{am_map}(x, doc = NULL, ...)
}
\arguments{
\item{x}{An Automerge map object}

\item{doc}{The document containing this object (automatically provided)}

\item{...}{Additional arguments (unused)}
}
\value{
Named list
}
\description{
Recursively converts an Automerge map to a named R list.
}
\keyword{internal}
