% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.list.am_text}
\alias{as.list.am_text}
\title{Convert Automerge text to character string}
\usage{
\method{as.list}{am_text}(x, doc = NULL, ...)
}
\arguments{
\item{x}{An Automerge text object}

\item{doc}{The document containing this object (automatically provided)}

\item{...}{Additional arguments (unused)}
}
\value{
Character string
}
\description{
Returns the text content as a character string.
}
\keyword{internal}
