% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automerge-package.R
\docType{package}
\name{automerge-package}
\alias{automerge}
\alias{automerge-package}
\title{automerge: R Bindings for 'Automerge' 'CRDT' Library}
\description{
Provides R bindings to the 'Automerge' Conflict-free Replicated Data Type ('CRDT') library. 'Automerge' enables automatic merging of concurrent changes without conflicts, making it ideal for distributed systems, collaborative applications, and offline-first architectures. The approach of local-first software was proposed in Kleppmann, M., Wiggins, A., van Hardenberg, P., McGranaghan, M. (2019) \doi{10.1145/3359591.3359737}. This package supports all 'Automerge' data types (maps, lists, text, counters) and provides both low-level and high-level synchronization protocols for seamless interoperability with 'JavaScript' and other 'Automerge' implementations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/posit-dev/automerge-r}
  \item \url{https://posit-dev.github.io/automerge-r/}
  \item Report bugs at \url{https://github.com/posit-dev/automerge-r/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Authors of the dependency Rust crates (see inst/AUTHORS file) [copyright holder]
}

}
\keyword{internal}
