% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length.am_object}
\alias{length.am_object}
\title{Get length of Automerge object}
\usage{
\method{length}{am_object}(x)
}
\arguments{
\item{x}{An Automerge object}
}
\value{
Integer length
}
\description{
Returns the number of elements/keys in an Automerge object.
}
