% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{names.am_map}
\alias{names.am_map}
\title{Get names from Automerge map object}
\usage{
\method{names}{am_map}(x)
}
\arguments{
\item{x}{An Automerge map object}
}
\value{
Character vector of key names
}
\description{
Returns the keys from a map object.
}
