% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_change}
\alias{print.am_change}
\title{Print Automerge change}
\usage{
\method{print}{am_change}(x, ...)
}
\arguments{
\item{x}{An Automerge change}

\item{...}{Additional arguments (unused)}
}
\value{
The change (invisibly)
}
\description{
Print Automerge change
}
\keyword{internal}
