% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_counter}
\alias{print.am_counter}
\title{Print Automerge counter}
\usage{
\method{print}{am_counter}(x, ...)
}
\arguments{
\item{x}{An Automerge counter}

\item{...}{Additional arguments (unused)}
}
\value{
The counter (invisibly)
}
\description{
Print Automerge counter
}
\keyword{internal}
