% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_cursor}
\alias{print.am_cursor}
\title{Print Automerge cursor}
\usage{
\method{print}{am_cursor}(x, ...)
}
\arguments{
\item{x}{An Automerge cursor}

\item{...}{Additional arguments (unused)}
}
\value{
The cursor (invisibly)
}
\description{
Print Automerge cursor
}
\keyword{internal}
