% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_list}
\alias{print.am_list}
\title{Print Automerge list object}
\usage{
\method{print}{am_list}(x, ...)
}
\arguments{
\item{x}{An Automerge list object}

\item{...}{Additional arguments (unused)}
}
\value{
The object (invisibly)
}
\description{
Print Automerge list object
}
\keyword{internal}
