% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_map}
\alias{print.am_map}
\title{Print Automerge map object}
\usage{
\method{print}{am_map}(x, ...)
}
\arguments{
\item{x}{An Automerge map object}

\item{...}{Additional arguments (unused)}
}
\value{
The object (invisibly)
}
\description{
Print Automerge map object
}
\keyword{internal}
