% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_object}
\alias{print.am_object}
\title{Print Automerge object (fallback for unknown types)}
\usage{
\method{print}{am_object}(x, ...)
}
\arguments{
\item{x}{An Automerge object}

\item{...}{Additional arguments (unused)}
}
\value{
The object (invisibly)
}
\description{
Print Automerge object (fallback for unknown types)
}
\keyword{internal}
