% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_syncstate}
\alias{print.am_syncstate}
\title{Print Automerge sync state}
\usage{
\method{print}{am_syncstate}(x, ...)
}
\arguments{
\item{x}{An Automerge sync state}

\item{...}{Additional arguments (unused)}
}
\value{
The sync state (invisibly)
}
\description{
Print Automerge sync state
}
\keyword{internal}
