% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_text}
\alias{print.am_text}
\title{Print Automerge text object}
\usage{
\method{print}{am_text}(x, ...)
}
\arguments{
\item{x}{An Automerge text object}

\item{...}{Additional arguments (unused)}
}
\value{
The object (invisibly)
}
\description{
Print Automerge text object
}
\keyword{internal}
