% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.am_uint64}
\alias{print.am_uint64}
\title{Print Automerge unsigned 64-bit integer}
\usage{
\method{print}{am_uint64}(x, ...)
}
\arguments{
\item{x}{An Automerge uint64}

\item{...}{Additional arguments (unused)}
}
\value{
The uint64 (invisibly)
}
\description{
Print Automerge unsigned 64-bit integer
}
\keyword{internal}
