% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{replace-am_object}
\alias{replace-am_object}
\alias{[[<-.am_object}
\alias{$<-.am_object}
\title{Replace in Automerge object}
\usage{
\method{[[}{am_object}(x, i) <- value

\method{$}{am_object}(x, name) <- value
}
\arguments{
\item{x}{An Automerge object}

\item{i}{Key name (character) for maps, or position (integer) for lists}

\item{value}{Value to store}

\item{name}{Key name (for \verb{$<-} operator, maps only)}
}
\value{
The object (invisibly)
}
\description{
Replace or insert values in an Automerge object using \verb{[[<-} or \verb{$<-}.
}
\examples{
doc <- am_create()

am_put(doc, AM_ROOT, "user", list(name = "Bob", age = 25L))
user <- am_get(doc, AM_ROOT, "user")
user

user[["name"]] <- "Alice"
user$age <- 30L

am_close(doc)

}
